+++ lib/src/components/loader/index.vue
@@ -6,7 +6,6 @@
 
 <script>
 export default {
-  name: 'common-loader-component',
   props: {
     show: Boolean
   }