+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 Changelog
 =========
 
+### 0.0.6 (2020-06-23)
+* Modal component added.
+
+
 ### 0.0.5 (2020-06-14)
 * Loader component added.
 * Textarea component added.
 * Input error classname was changed.
-* Input required prop added.
+* Input "required" prop added.
 
 
 ### 0.0.4 (2020-06-13)